+++ src/world/mapping.rs
@@ -10,11 +10,11 @@ use crate::rrl_math::Position;
 
 pub trait Mapping
 {
-    fn height(&self) -> usize;
+    fn height(&self) -> u32;
 
-    fn width(&self) -> usize;
+    fn width(&self) -> u32;
     
-    fn bounds(&self) -> ( usize, usize)
+    fn bounds(&self) -> ( u32, u32)
     {
         ( self.width(), self.height() )
     }
@@ -24,7 +24,7 @@ pub trait Mapping
         self.is_i32_in_height( pos_y ) && self.is_i32_in_width( pos_x )
     }
 
-    fn is_in_bounds(&self, pos_x: usize, pos_y: usize) -> bool
+    fn is_in_bounds(&self, pos_x: u32, pos_y: u32) -> bool
     {
         self.is_in_height( pos_y ) && self.is_in_width( pos_x )
     }
@@ -36,20 +36,20 @@ pub trait Mapping
 
     fn is_i32_in_height(&self, pos_y: i32) -> bool
     {
-        pos_y >= 0 && self.is_in_height( pos_y as usize )
+        pos_y >= 0 && self.is_in_height( pos_y as u32 )
     }
 
     fn is_i32_in_width(&self, pos_x: i32) -> bool
     {
-        pos_x >= 0 && self.is_in_width( pos_x as usize )
+        pos_x >= 0 && self.is_in_width( pos_x as u32 )
     }
 
-    fn is_in_height(&self, pos_y: usize) -> bool
+    fn is_in_height(&self, pos_y: u32) -> bool
     {
         pos_y < self.height()
     }
 
-    fn is_in_width(&self, pos_x: usize) -> bool
+    fn is_in_width(&self, pos_x: u32) -> bool
     {
         pos_x < self.width()
     }
